Move federationout api content
